refactor(kyc): move non-component badge helpers out of KycStatusIcon
Fast-refresh rule (only-export-components) flags non-component exports
from component files; the new KYC_BADGE_BASE/getKycBadgeLabel/
kycBadgeColors exports tripped the quality-gate baseline. Move them,
plus the pre-existing getEffectiveKycStatus and the status config map,
into kyc-status-badge.helpers.ts.
