+++ app/community/[communityId]/(whitelabel)/applications/[applicationId]/components/ApplicationKycCard.tsx
@@ -1,6 +1,7 @@
 "use client";
 
-import { getEffectiveKycStatus, KycStatusBadge } from "@/components/KycStatusIcon";
+import { KycStatusBadge } from "@/components/KycStatusIcon";
+import { getEffectiveKycStatus } from "@/components/kyc-status-badge.helpers";
 import { useKycConfig, useKycStatus } from "@/hooks/useKycStatus";
 import { kycStatusDescriptions } from "@/src/features/kyc/lib/status-config";
